amiaterm: WIP for using IOF_QUICK + large block reads
This is a WIP for me to experiment with NOT doing lots of single byte
driver calls for serial IO but doing the larger block receives
as blocks.  At 19200 I can now receive at 1400 bytes per second over
xmodem without the Amiga 500 CPU actually hitting 100% usage.
38400 starts erroring out and then the code ends up dying -
unfortunately the xmodem code (and amigaterm in general) isn't doing
timeouts at all yet so it's not aborting something like a short
128 byte read and retrying the block.

The hack is a bit hacky because of how the default single byte
read scheduling is happening.

anyway:

* track whether read IO is scheduled or aborted.  I'm not currently
  tracking when it's actually completed; it's not yet needed here
  but it would be "correct" to do so.
* abort any scheduled read IO when we wrap up the program for exit.
* convert the xmodem read path to call readchar_buf() for the whole
  128 byte buffer, and calculate a checksum over that.
* Implement both a naive readchar_buf() using single byte readchar()'s
  like I'm currently doing, and a hack to read the first byte without
  scheduling a follow-up IO, then len-1 as the second transaction.

It's dirty as an experiment but with all of this work I'm now hitting
1400 bytes rather than 600-800 bytes at 19200 serial between my PC
and the Amiga 500 (receiving.)  Transmit would likely benefit too
from a similar hack but that can come later.

In any case, it's proving that things are heading in the right direction.
Next up will be cleaning up the xmodem receive code and readchar/readchar_buf
to use timeouts (well, and I'll need to likely figure out how to schedule
and cancel timer signals!) to retry stuff, and then re-do how read IO
is done so I can do the whole 128 byte read in one hit.

(It'd sure be nice to have 38400 working!)