amigaterm: more figuring out what the heck is going on with the serial error path.
I still don't know what to do when I hit a waitio error.

The signal in the flush or receive loop claims the serial port is ready
but checking it returns "none available" until an event receive happens.
So maybe I need to abort at this point? I'm not sure.