[amigaterm] abstract out the UI keypress check in the serial read functions
Instead of directly poking at the UI path for keypress messages,
use a pair of functions that provide the signal to wait on and a check
callback.

That way we can compile in a CLI abort handler later on, and the
read code doesn't need to know about the UI stuff.

So far so good; now the timeout/aborted global nonsense.  Ugh.