[amigaterm] migrate the screen emit stuff out
It turns out the per-character writing is pretty expensive; a bunch of it
is the hoops jumped through for the cursor updates.

So let's start refactoring this out; refactoring out the cursor update
will be next so bulk strings and large amounts of terminal echoing can
be done without updating the cursor on every character.