[amigaterm] Start accelerating the emit/emits functions.
"accelerating" is ... likely the wrong word here.  Drawing the cursor
after /each/ character being echoed is super slow.  For emits(), if
this is delayed then the bulk wirtes happen /much/ faster.

It's easier to accelerate for emits() because the cursor can just
be echoed at the end of the burst.

But for individual character echoes it's a bit less easy right now
because we don't know when a "good" time to update the cursor
would actually be.  I mean honestly I can't imagine drawing that
cursor should be THAT expensive, but this is a 7MHz 68000.

The #if 1 bits are around the cursor echoing so I can tinker with
it to measure timing.

Anyway, it's mostly a big no-op.  The screen location stuff is now
a global inside this source file so it can be used in multiple
places; I'll go clean it up and stick it behind a macro/inline
function.