[amigaterm] Refactor cursor drawing / next line logic
* refactor out the "next line?" logic so ...
* ... I can then use draw_cursor()
* remove the cursor display from _emit()
* change emits() to use _emit(), so now it's much faster.

This is much, much faster when printing out the bulk text via
emits(), but it's slow via emit() because of the double cursor
display stuff.

This is also likely why it's just wrong on the 1200 - the different
display depth means the xor operation isn't behaving the same and
clearing the cursor.  I'll have to go and tinker some more with this.