[amigaterm] refactor the screen drawing stuff to use intuition/font sizes
This removes the bulk of the hard-coded size assumptions about fonts
and screen gadgets.  There's a couple hard-coded places left that
I've added XXX'es to, and there's some left over cursor trash left
on the screen that i need to clean up.

However, this actually works.  It works well enough to scroll, wrap,
and work in both 60 and 80 character system font modes on Workbench 1.3.

I'll clean up the rest of this stuff in a subsequent commit and then
test it on Workbench 3.1 on the Amiga 1200 to see if I can (a) use
the higher resolutions there, and (b) the cursor drawing / scrolling
trash issues are gone.