amigaterm: mess around with when we draw cursors
drawing the cursor seems SUPER slow.  I don't know if it's as slow
as say, drawing a single character, but during normal emit() we're
effectively rendering three things - xor/delete cursor, character,
then draw cursor.

For bulk serial receives, we likely don't care about the rendering
of the cursor on the screen as we're about to delete it.  So as
an experiment, only draw the cursor if we're waiting for input.

This drastically improves the speed of rendering character by character
receive and render.  It's in line with the bulk emit string routine
which only echoes the cursor change at the start and end of the
rendered string.

This isn't complete - I'm not tracking when I drew the cursor
and need to /clear/ the cursor; that'll require some tactical tracking
of the cursor state.  I'll do it when it's not midnight.

(Also yes it would be much faster if I would just use a sprite
as a cursor, but I'm lazy right now and want to see how far I can
push this stuff before I start looking at optionally supporting
more of ANSI, colours, etc.)